refactor: RaceEntry로 클래스명 수정
