+++ src/domain/index.js
@@ -1,4 +1,4 @@
 export { default as Race } from "./Race.js";
 export { default as Racer } from "./Racer.js";
 export { default as RaceScoreboard } from "./RaceScoreboard.js";
-export { default as RacerRegistry } from "./RacerRegistry.js";
+export { default as RaceEntry } from "./RaceEntry.js";
